+++ src/globals.ts
@@ -3,7 +3,6 @@ export type webOSLaunchParams = Record<string, unknown>;
 declare global {
   interface Window {
     launchParams?: string | webOSLaunchParams;
-    __ytaf_debug__?: boolean;
     navigate?: (direction: 'left' | 'right' | 'up' | 'down') => void;
     __spatialNavigation__?: { keyMode: string };
     ytaf_showOptionsPanel?: (visible?: boolean) => void;